extracted main program to program.py file and added OTA functionality
