+++ boot.py
@@ -1,0 +1,1 @@
+updates_available = False
